Split functions into separate modules for implementation and wrapper
