+++ modal/experimental.py
@@ -5,12 +5,12 @@ from typing import Any, Callable, Optional
 from modal_proto import api_pb2
 
 from ._clustered_functions import ClusterInfo, get_cluster_info as _get_cluster_info
+from ._functions import _Function
 from ._object import _get_environment_name
 from ._runtime.container_io_manager import _ContainerIOManager
 from ._utils.async_utils import synchronizer
 from .client import _Client
 from .exception import InvalidError
-from .functions import _Function
 from .partial_function import _PartialFunction, _PartialFunctionFlags
